# docs/OpenApi.yaml
openapi: 3.0.3
info:
  title: io-dono Backend API
  description: >
    Specifica OpenAPI 3 completa delle API RESTful del backend io-dono.
    Include autenticazione locale/Google, gestione donazioni e segnalazioni.
  version: 1.0.0
  contact:
    name: io-dono API Team
    email: support@io-dono.local
  license:
    name: Proprietary
servers:
  - url: http://localhost:3000
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /api/auth/users:
    post:
      tags: [Auth]
      summary: Registrazione utente locale
      description: >
        Crea un nuovo utente locale con ruolo applicativo.
        Se l'email è già presente il sistema risponde con errore 400.
      operationId: registerUser
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRegisterRequest'
            examples:
              validDonor:
                $ref: '#/components/examples/AuthRegisterRequestExample'
      responses:
        '201':
          description: Utente creato e token JWT emesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokenUserResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /api/auth/sessions:
    post:
      tags: [Auth]
      summary: Login locale
      description: >
        Autentica un utente locale tramite email/password.
        Per account creati con Google OAuth il login locale è bloccato.
      operationId: createSession
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthLoginRequest'
      responses:
        '200':
          description: Login riuscito
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokenUserResponse'
              examples:
                success:
                  $ref: '#/components/examples/AuthLoginResponseExample'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags: [Auth]
      summary: Logout esplicito (invalidate client token)
      operationId: deleteSession
      responses:
        '204':
          description: Sessione terminata
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /api/auth/google/sessions:
    post:
      tags: [Auth]
      summary: Exchange token Google
      description: >
        Scambia il token Google (`token`) con:
        - `loginToken` se l'utente esiste
        - `registrationToken` se deve completare la registrazione
      operationId: exchangeGoogleToken
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GoogleTokenExchangeRequest'
      responses:
        '200':
          description: Login token emesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GoogleExchangeLoginResponse'
        '201':
          description: Registration token emesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GoogleExchangeRegistrationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/auth/google/users:
    post:
      tags: [Auth]
      summary: Completa registrazione utente Google
      description: >
        Richiede Authorization Bearer con registration token (type=registration),
        non un login JWT standard.
      operationId: registerGoogleUser
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GoogleRegisterRequest'
      responses:
        '201':
          description: Utente Google creato e token login emesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokenUserResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    BadRequest:
      description: Richiesta non valida
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Non autenticato / token non valido
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    InternalServerError:
      description: Errore interno server
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ObjectId:
      type: string
      pattern: '^[a-fA-F0-9]{24}$'
      example: 65f3c2b9f4bbf8a7d3ea0011
    ErrorResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
          example: Errore del server
        error:
          type: string
          nullable: true
    UserProfile:
      type: object
      properties:
        donorType:
          type: string
          enum: [PRIVATE, COMMERCIAL]
        commercialHours:
          type: string
    UserPublic:
      type: object
      properties:
        _id:
          $ref: '#/components/schemas/ObjectId'
        email:
          type: string
          format: email
        googleId:
          type: string
          nullable: true
        role:
          type: string
          enum: [DONOR, ASSOCIATION, ADMIN]
        name:
          type: string
        phoneNumber:
          type: string
        address:
          type: string
        profile:
          $ref: '#/components/schemas/UserProfile'
        solidarityPoints:
          type: integer
          default: 0
        redeemedRewards:
          type: array
          items:
            $ref: '#/components/schemas/ObjectId'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    AuthRegisterRequest:
      type: object
      required: [email, password, name, role, phoneNumber, address]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        name:
          type: string
        role:
          type: string
          enum: [DONOR, ASSOCIATION, ADMIN]
        phoneNumber:
          type: string
        address:
          type: string
        profile:
          $ref: '#/components/schemas/UserProfile'
    AuthLoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    AuthTokenUserResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
          description: JWT login token
        user:
          $ref: '#/components/schemas/UserPublic'
...
    GoogleTokenExchangeRequest:
      type: object
      required: [token]
      properties:
        token:
          type: string
          description: Google ID token inviato dal frontend
    GoogleExchangeLoginResponse:
      type: object
      required: [loginToken, user]
      properties:
        loginToken:
          type: string
          description: JWT login token applicativo
        user:
          $ref: '#/components/schemas/UserPublic'
    GoogleExchangeRegistrationResponse:
      type: object
      required: [registrationToken]
      properties:
        registrationToken:
          type: string
          description: JWT temporaneo per completamento registrazione Google
    GoogleRegisterRequest:
      type: object
      required: [role, phoneNumber, address]
      properties:
        role:
          type: string
          enum: [DONOR, ASSOCIATION, ADMIN]
        phoneNumber:
          type: string
        address:
          type: string
        profile:
          $ref: '#/components/schemas/UserProfile'
...
  examples:
    AuthRegisterRequestExample:
      summary: Registrazione donatore
      value:
        email: donor@example.com
        password: Password123!
        name: Mario Rossi
        role: DONOR
        phoneNumber: "3331234567"
        address: Via Roma 1, Trento
        profile:
          donorType: PRIVATE
    AuthLoginResponseExample:
      summary: Login riuscito
      value:
        token: eyJhbGciOi...
        user:
          _id: 65f3c2b9f4bbf8a7d3ea0011
          email: donor@example.com
          role: DONOR
          name: Mario Rossi
          phoneNumber: "3331234567"
          address: Via Roma 1, Trento
          solidarityPoints: 0
...